+++ warp/codegen.py
@@ -1205,7 +1205,6 @@ cpu_module_header = '''
 #define int(x) cast_int(x)
 #define adj_int(x, adj_x, adj_ret) adj_cast_int(x, adj_x, adj_ret)
 
-
 using namespace wp;
 
 '''