+++ warp/tape.py
@@ -10,14 +10,13 @@ import warp as wp
 
 class Tape:
 
-    def __init__(self, capture=False):
+    def __init__(self):
 
         self.gradients = {}
         self.launches = []
 
-        self.capture = capture
-        self.capture_graph_forward = None
-        self.capture_graph_backward = None
+        self.loss = None
+
 
     def __enter__(self):      
         if (wp.context.runtime.tape != None):
@@ -25,17 +24,11 @@ class Tape:
 
         wp.context.runtime.tape = self
 
-        if self.capture:
-            wp.capture_begin()
-
     def __exit__(self, exc_type, exc_value, traceback):
         if (wp.context.runtime.tape == None):
             raise RuntimeError("Warp: Error, ended tape capture, but tape not present")            
 
         wp.context.runtime.tape = None
-
-        if self.capture:
-            self.capture_graph_forward = wp.capture_end()
 
     # adj_outputs is a mapping from output tensor -> adjoint of the output
     # after running backward the gradients of tensors may be retrieved by:
@@ -44,38 +37,25 @@ class Tape:
     #
     def backward(self, loss: wp.array=None, grads: dict=None):
 
-        # if scalar loss is specified then allocate 
+        # if scalar loss is specified then initialize
         # a 'seed' array for it, with gradient of one
-        if (loss):
+        if loss:
             
-            if len(loss) > 1:
+            if loss.size > 1 or wp.types.type_length(loss.dtype) > 1:
                 raise RuntimeError("Can only return gradients for scalar loss functions.")
 
-            self.gradients[loss] = wp.array(np.ones(1), dtype=wp.float32, device=loss.device)
+            if loss.requires_grad == False:
+                raise RuntimeError("Scalar loss arrays should have requires_grad=True set before calling Tape.backward()")
 
-        # force allocation of all adjoints before capture
-        if self.capture:
-
-            for launch in reversed(self.launches):
-                inputs = launch[2]
-                outputs = launch[3]
-
-                for a in inputs:
-                    self.get_adjoint(a)
-        
-                for a in outputs:
-                    self.get_adjoint(a)
-
-            wp.capture_begin()
-
-            # zero all gradients (except for loss grad)
-            for k, v in self.gradients.items():
-                if (k != loss):
-                    v.zero_()
-
-        # insert any user specified gradients (e.g.: from Torch)
-        if (grads):
-            self.gradients.update(grads)
+            # set the seed grad to 1.0
+            loss.grad.fill_(1.0)
+            
+        # simply apply dict grads to objects
+        # this is just for backward compat. with
+        # existing code before we added wp.array.grad attribute
+        if grads:
+            for a, g in grads.items():
+                a.grad = g
 
         # run launches backwards
         for launch in reversed(self.launches):
@@ -108,44 +88,26 @@ class Tape:
                 adjoint=True)
 
 
-        if self.capture:
-            self.capture_graph_backward = wp.capture_end()
-            
-
     # record a kernel launch on the tape
     def record(self, kernel, dim, inputs, outputs, device):
         self.launches.append([kernel, dim, inputs, outputs, device])
 
 
-    # returns the adjoint version of a tensor used in the computation
+    # returns the adjoint of a kernel parameter
     def get_adjoint(self, a):
         
         if isinstance(a, wp.array) == False:
-            # if input is a simple type (e.g.: float, vec3, etc) just return a value copy
+            # if input is a simple type (e.g.: float, vec3, etc) then
+            # no gradient needed (we only return gradients through arrays)
             return a
-
-        elif a in self.gradients:
-            # try and find adjoint array in map
-            return self.gradients[a]
-                    
-        elif wp.types.type_is_int(a.dtype) or a.requires_grad == False:
-            # otherwise if input is an array that is integer typed or doesn't require grad then return null array
-            return None
-
+           
         else:
-            # otherwise allocate a zero array for the array adjoint
-            adj = wp.zeros_like(a)
-            self.gradients[a] = adj
-            return adj
-
-    def replay(self):
-        
-        if self.capture == False:
-            raise RuntimeError("Cannot replay from a non-captured Tape")
-
-        wp.capture_launch(self.capture_graph_forward)
-        wp.capture_launch(self.capture_graph_backward)
-
+            # keep track of all gradients used by the tape (for zeroing)
+            # ignore the scalar loss since we don't want to clear its grad
+            if a.grad:
+                self.gradients[a] = a.grad
+                
+            return a.grad
 
     def reset(self):
         
@@ -156,3 +118,4 @@ class Tape:
 
         for a in self.gradients.values():
             a.zero_()
+